refactor: consolidate routes
