+++ src/routes/v1/authenticatedSites/images.js
