+++ src/routes/v1/authenticatedSites/index.js
@@ -1,20 +1,20 @@
 const express = require("express")
 
-const collectionPagesRouter = require("@routes/authenticatedSites/collectionPages")
-const collectionsRouter = require("@routes/authenticatedSites/collections")
-const directoryRouter = require("@routes/authenticatedSites/directory")
-const documentsRouter = require("@routes/authenticatedSites/documents")
-const foldersRouter = require("@routes/authenticatedSites/folders")
-const homepageRouter = require("@routes/authenticatedSites/homepage")
-const imagesRouter = require("@routes/authenticatedSites/images")
-const mediaSubfolderRouter = require("@routes/authenticatedSites/mediaSubfolder")
-const navigationRouter = require("@routes/authenticatedSites/navigation")
-const netlifyTomlRouter = require("@routes/authenticatedSites/netlifyToml")
-const pagesRouter = require("@routes/authenticatedSites/pages")
-const resourcePagesRouter = require("@routes/authenticatedSites/resourcePages")
-const resourceRoomRouter = require("@routes/authenticatedSites/resourceRoom")
-const resourcesRouter = require("@routes/authenticatedSites/resources")
-const settingsRouter = require("@routes/authenticatedSites/settings")
+const collectionPagesRouter = require("@routes/v1/authenticatedSites/collectionPages")
+const collectionsRouter = require("@routes/v1/authenticatedSites/collections")
+const directoryRouter = require("@routes/v1/authenticatedSites/directory")
+const documentsRouter = require("@routes/v1/authenticatedSites/documents")
+const foldersRouter = require("@routes/v1/authenticatedSites/folders")
+const homepageRouter = require("@routes/v1/authenticatedSites/homepage")
+const imagesRouter = require("@routes/v1/authenticatedSites/images")
+const mediaSubfolderRouter = require("@routes/v1/authenticatedSites/mediaSubfolder")
+const navigationRouter = require("@routes/v1/authenticatedSites/navigation")
+const netlifyTomlRouter = require("@routes/v1/authenticatedSites/netlifyToml")
+const pagesRouter = require("@routes/v1/authenticatedSites/pages")
+const resourcePagesRouter = require("@routes/v1/authenticatedSites/resourcePages")
+const resourceRoomRouter = require("@routes/v1/authenticatedSites/resourceRoom")
+const resourcesRouter = require("@routes/v1/authenticatedSites/resources")
+const settingsRouter = require("@routes/v1/authenticatedSites/settings")
 
 const getAuthenticatedSitesSubrouter = ({ authMiddleware }) => {
   const authenticatedSitesSubrouter = express.Router({ mergeParams: true })
