+++ src/routes/v1/authenticatedSites/pages.js
@@ -9,9 +9,9 @@ const {
 } = require("@middleware/routeHandler")
 
 // Import classes
-const { Collection } = require("@classes/Collection.js")
+const { Collection } = require("@classes/Collection")
 const { CollectionConfig } = require("@classes/Config")
-const { File, PageType, CollectionPageType } = require("@classes/File.js")
+const { File, PageType, CollectionPageType } = require("@classes/File")
 const { Subfolder } = require("@classes/Subfolder")
 
 const { deslugifyCollectionName } = require("@utils/utils")
