+++ src/routes/v1/authenticatedSites/resources.js
