+++ src/routes/v2/authenticatedSites/resourceCategories.js
