+++ src/routes/v2/authenticatedSites/resourcePages.js
